feat: Add updateUser function
This commit adds a new file `src/endpoints/user/updateUser.ts` that exports the `updateUser` function. This function is used to update user information by making a POST request to the API endpoint `/user`. It takes an authentication token and a payload as parameters and returns a `MeResponse` object.

Note: The commit message follows the established convention of using a prefix (`feat`) to indicate a new feature or functionality being added.